添加sse，删除了一些字段
