+++ modules/stun/stun.go
@@ -1,391 +1,520 @@
 package stun
 
-// import (
-// 	"context"
-// 	"fmt"
-// 	"linkstar/global"
-// 	"linkstar/modules/stun/model"
-// 	"net"
-// 	"strings"
-// 	"time"
+import (
+	"context"
+	"fmt"
+	"net"
+	"strings"
+	"time"
 
-// 	"github.com/libp2p/go-reuseport"
-// 	"github.com/pion/stun"
-// 	"github.com/sirupsen/logrus"
-// )
+	"github.com/libp2p/go-reuseport"
+	"github.com/pion/stun"
+	"github.com/sirupsen/logrus"
+)
 
-// // RunStunTunnelWithContext 实现内网穿透逻辑  支持 context 取消的穿透逻辑
-// func RunStunTunnelWithContext(ctx context.Context, targetIP string, service *model.Service) error {
-// 	protocol := strings.ToLower(service.Protocol) // 转为小写
+// STUN 内网穿透实现
+func (STUNRunner) Run(ctx context.Context, req STUNRequest, onState func(STUNState)) error {
 
-// 	localAddr := fmt.Sprintf("%s:0", global.StunConfig.LocalIP) //端口为0任意端口
+	// 验证环境，确保不缺东西
+	protocol, err := formatProtocol(req.Protocol)
+	if err != nil {
+		return err
+	}
+	localIP := Runtime.Network.LocalIP
+	if localIP == "" {
+		return fmt.Errorf("get local IP failed")
+	}
+	stunServer, err := Runtime.STUNService.GetBestSTUNServer()
+	if err != nil {
+		stunServer, err = Runtime.STUNService.GetBackupSTUNServer()
+		if err != nil {
+			return err
+		}
+	}
 
-// 	// STUN 拨号
-// 	stunConn, err := reuseport.Dial(protocol, localAddr, global.StunConfig.BestSTUN)
-// 	if err != nil {
-// 		return fmt.Errorf("STUN拨号失败 [%s]:%w", global.StunConfig.BestSTUN, err)
-// 	}
+	// 端口复用连接STUN服务器
+	localAddr := fmt.Sprintf("%s:0", localIP)
+	stunConn, err := reuseport.Dial(protocol, localAddr, stunServer)
+	if err != nil {
+		primaryErr := err
+		primarySTUNServer := stunServer
+		// 可能 best STUN 掉线了，使用备用的
+		stunServer, err = Runtime.STUNService.GetBackupSTUNServer()
+		if err != nil {
+			return fmt.Errorf("connect stun server %s failed: %v; get backup stun server: %w", primarySTUNServer, primaryErr, err)
+		}
+		// 再次发起连接
+		stunConn, err = reuseport.Dial(protocol, localAddr, stunServer)
+		if err != nil {
+			return fmt.Errorf("connect stun server %s failed: %v; connect backup stun server %s: %w", primarySTUNServer, primaryErr, stunServer, err)
+		}
+	}
+	defer stunConn.Close()
 
-// 	var localPort uint16
-// 	if protocol == "tcp" {
-// 		localPort = uint16(stunConn.LocalAddr().(*net.TCPAddr).Port)
-// 	} else {
-// 		localPort = uint16(stunConn.LocalAddr().(*net.UDPAddr).Port)
-// 	}
+	//获取本地端口
+	var localPort uint16
+	switch protocol {
+	case "tcp":
+		localPort = uint16(stunConn.LocalAddr().(*net.TCPAddr).Port)
+	case "udp":
+		localPort = uint16(stunConn.LocalAddr().(*net.UDPAddr).Port)
+	}
 
-// 	// STUN 握手
-// 	var publicIP string
-// 	var publicPort int
-// 	if protocol == "tcp" {
-// 		publicIP, publicPort, err = doTcpStunHandshake(stunConn)
-// 	} else {
-// 		udpConn := stunConn.(*net.UDPConn)
-// 		stunServerAddr, _ := net.ResolveUDPAddr("udp", global.StunConfig.BestSTUN)
-// 		publicIP, publicPort, err = doUDPStunHandshake(udpConn, stunServerAddr)
-// 	}
-// 	if err != nil {
-// 		stunConn.Close()
-// 		return fmt.Errorf("与STUN服务器握手失败:%w", err)
-// 	}
+	// 发送STUN 请求,获取外部端口
+	publicIP, publicPort, err := doStunHandshake(stunConn)
+	if err != nil {
+		return fmt.Errorf("stun handshake: %w", err)
+	}
+	if protocol == "udp" {
+		logrus.Infof("UDP:%s:%d", publicIP, publicPort)
+	}
+	onState(STUNState{State: STUNLog, Log: fmt.Sprintf("通过STUN获取外部信息：%s %s:%d", protocol, publicIP, publicPort)})
 
-// 	// 端口复用监听
-// 	listenAddr := fmt.Sprintf("%s:%d", global.StunConfig.LocalIP, localPort)
-// 	listener, err := reuseport.Listen(protocol, listenAddr) // 使用reuseport SO_REUSEPORT 可以复用端口
-// 	if err != nil {
-// 		stunConn.Close()
-// 		return fmt.Errorf("端口监听失败：%w", err)
-// 	}
+	// 在本地端口复用监听，接受穿透后的入站连接
+	var tcpListener net.Listener
+	var udpListener *net.UDPConn
+	listenAddr := fmt.Sprintf("%s:%d", localIP, localPort)
+	if protocol == "tcp" {
 
-// 	// 路由器upnp映射
-// 	upnpCtx, upnpCancel := context.WithTimeout(ctx, 25*time.Second) //创建upnp的ctx
-// 	defer upnpCancel()
-// 	description := fmt.Sprintf("LinkStar-%s", service.Name)
-// 	err = AddPortMappingQueue(upnpCtx, localPort, localPort, "TCP", description)
-// 	if err != nil {
-// 		logrus.Warnf("[%s] UPnP 映射失败 (非致命): %v", service.Name, err)
-// 		// todo 处理失败
-// 	} else {
-// 		logrus.Infof("[%s] UPnP 映射成功: 路由器 WAN:%d -> 本机:%d", service.Name, localPort, localPort)
-// 	}
+		tcpListener, err = reuseport.Listen(protocol, listenAddr)
+		if err != nil {
+			return fmt.Errorf("listen on %s failed: %w", listenAddr, err)
+		}
+		defer tcpListener.Close()
+	} else {
+		// UDP 没有 Listener 概念，用 ListenUDP 接收入站包
+		var udpListenAddr *net.UDPAddr
+		udpListenAddr, resolveErr := net.ResolveUDPAddr("udp", listenAddr)
+		if resolveErr != nil {
+			return fmt.Errorf("resolve udp addr %s failed: %w", listenAddr, resolveErr)
+		}
+		// ↓ 换成 reuseport.ListenPacket，和 stunConn 共用同一个端口
+		packetConn, listenErr := reuseport.ListenPacket("udp", udpListenAddr.String())
+		if listenErr != nil {
+			return fmt.Errorf("listen udp on %s failed: %w", listenAddr, listenErr)
+		}
+		udpListener = packetConn.(*net.UDPConn)
+		defer udpListener.Close()
 
-// 	// ctx 取消时关闭连接，让所有阻塞调用立即返回
-// 	// go func() {
-// 	// 	<-ctx.Done()
-// 	// 	logrus.Infof("[%s] ctx 取消，关闭连接", service.Name)
-// 	// 	stunConn.Close()
-// 	// 	listener.Close()
-// 	// }()
-// 	// 确保所有子 goroutine（健康检查、Accept循环）能感知到退出信号，不再泄露。
-// 	innerCtx, innerCancel := context.WithCancel(ctx)
-// 	defer innerCancel()
+	}
 
-// 	defer func() {
-// 		logrus.Infof("[%s] 正在清理资源...", service.Name)
-// 		stunConn.Close()
-// 		listener.Close()
-// 		go DeletePortMapping(localPort, protocol)
-// 		service.PunchSuccess = false
-// 		service.ExternalPort = 0
-// 	}()
+	// UPNP启动 映射 网关的localPort->localip:localPort
+	if req.UseUPnP {
+		upnpCtx, upnpCancel := context.WithTimeout(ctx, 20*time.Second)
+		err := AddPortMappingByQueueWithLocalIP(
+			upnpCtx,
+			localPort,
+			localPort,
+			strings.ToUpper(protocol),
+			fmt.Sprintf("EasyLink-%s", req.ServiceName),
+			localIP,
+		)
+		if err != nil {
+			upnpCancel()
+			return err
+		}
+		// 释放资源
+		upnpCancel()
+		defer func() {
+			go DeletePortMapping(localPort, strings.ToUpper(protocol))
+		}()
+	}
+	onState(STUNState{State: STUNLog, Log: fmt.Sprintf("UPNP启动成功：路由:%d->%s:%d", localPort, localIP, localPort)})
 
-// 	errCh := make(chan error, 3)
-// 	logrus.Infof("%v %v %v", localPort, publicIP, publicPort)
+	// 把端口传出去
+	if onState != nil {
+		onState(STUNState{State: STUNMapped, ExternalPort: uint16(publicPort)}) //todo统一一下命名
+	}
 
-// 	// 存储数据
-// 	var publicURL string
-// 	if service.TLS {
-// 		publicURL = fmt.Sprintf("https://%s:%d", publicIP, publicPort)
-// 	} else {
-// 		publicURL = fmt.Sprintf("http://%s:%d", publicIP, publicPort)
-// 	}
+	// 保活
+	innerCtx, innerCancel := context.WithCancel(ctx)
+	errCh := make(chan error, 2)
+	defer innerCancel()
+	if protocol == "tcp" { //tcp 保活
+		go func() {
+			if err := tcpStunHealthKeepAlive(
+				innerCtx,
+				stunConn,
+				publicIP,
+				publicPort,
+				localPort,
+				localIP,
+				func() {
+					if onState != nil {
+						onState(STUNState{State: STUNAlive, ExternalPort: uint16(publicPort)})
+					}
+				},
+			); err != nil && ctx.Err() == nil {
+				errCh <- fmt.Errorf("tcp keepalive failed: %w", err)
+			}
+		}()
+	} else if protocol == "udp" { //udp 保活
+		go func() {
+			if err := udpStunHealthKeepAlive(
+				innerCtx,
+				stunConn,
+				// udpListener, // 用来收包
+				publicIP,
+				publicPort,
+				localPort,
+				localIP,
+				func() {
+					if onState != nil {
+						onState(STUNState{State: STUNAlive, ExternalPort: uint16(publicPort)})
+					}
+				},
+			); err != nil && ctx.Err() == nil {
+				errCh <- fmt.Errorf("udp keepalive failed: %w", err)
+			}
+		}()
+	}
 
-// 	service.ExternalPort = uint16(publicPort)
-// 	service.PunchSuccess = true
+	// 接收外网连接，转发到对应内网服务
+	go func() {
+		targetAddr := fmt.Sprintf("%s:%d", req.TargetIP, req.InternalPort) // 构建目标ip+端口
+		if protocol == "tcp" {
+			for {
+				clientConn, err := tcpListener.Accept()
+				if err != nil {
+					if ctx.Err() != nil {
+						errCh <- nil
+					} else {
+						errCh <- fmt.Errorf("listener closed unexpectedly: %w", err)
+					}
+					return
+				}
+				go ForwardTCP(clientConn, targetAddr, protocol) //转发到对应的内网服务
+			}
+		} else {
+			buf := make([]byte, 65535)
+			for {
+				n, remoteAddr, err := udpListener.ReadFromUDP(buf)
+				if err != nil {
+					if ctx.Err() != nil {
+						errCh <- nil
+					} else {
+						errCh <- fmt.Errorf("udp listener closed: %w", err)
+					}
+					return
+				}
+				data := make([]byte, n)
+				copy(data, buf[:n])
+				go ForwardUDP(udpListener, remoteAddr, data, targetAddr)
+			}
+		}
 
-// 	// 开启保活
-// 	if protocol == "tcp" {
-// 		go func() {
-// 			err = tcpStunHealthCheck(innerCtx, stunConn, publicIP, publicPort, localPort, service)
-// 			if err != nil {
-// 				service.PunchSuccess = false
-// 				errCh <- fmt.Errorf("TCP健康检查失败: %w", err)
-// 			}
-// 		}()
-// 	} else {
-// 		go func() {
-// 			udpConn := stunConn.(*net.UDPConn)
-// 			stunServerAddr, _ := net.ResolveUDPAddr("udp", global.StunConfig.BestSTUN)
-// 			err = udpStunHealthCheck(udpConn, stunServerAddr, publicPort, localPort)
-// 			if err != nil {
-// 				service.PunchSuccess = false
-// 				errCh <- fmt.Errorf("UDP健康检查失败: %w", err)
-// 			}
-// 		}()
-// 	}
+	}()
 
-// 	go func() {
-// 		targetAddr := fmt.Sprintf("%s:%d", targetIP, service.InternalPort)
-// 		for {
-// 			clientConn, err := listener.Accept()
-// 			if err != nil {
-// 				errCh <- fmt.Errorf("监听器退出: %w", err)
-// 				return
-// 			}
-// 			logrus.Infof("[%s] 收到外部连接: %s", service.Name, clientConn.RemoteAddr())
-// 			go Forward(clientConn, targetAddr, protocol)
-// 		}
-// 	}()
+	return <-errCh
+}
 
-// 	// 输出访问数据
-// 	logrus.Infof("   访问地址: %s", publicURL)
-// 	return <-errCh
-// }
+// ===================== 内部 函数 ===============
 
-// // 与STUN服务器握手TCP
-// func doTcpStunHandshake(conn net.Conn) (string, int, error) {
+// 格式化 Protocol 输入
+func formatProtocol(protocol string) (string, error) {
+	value := strings.ToLower(strings.TrimSpace(protocol)) // 去除空格转小写
+	if value == "" {
+		value = "tcp"
+	}
+	switch value {
+	case "tcp", "udp":
+		return value, nil
+	default:
+		return "", fmt.Errorf("unsupported protocol: %s", protocol)
+	}
 
-// 	// 发送STUN请求
-// 	msg := stun.MustBuild(stun.TransactionID, stun.BindingRequest)
-// 	if _, err := conn.Write(msg.Raw); err != nil {
-// 		return "", 0, fmt.Errorf("发送STUN请求失败%s", err)
-// 	}
+}
 
-// 	// 读取响应
-// 	conn.SetDeadline(time.Now().Add(3 * time.Second))
-// 	defer conn.SetReadDeadline(time.Time{})
+// 发送 STUN 请求
+func doStunHandshake(conn net.Conn) (string, int, error) {
+	msg := stun.MustBuild(stun.TransactionID, stun.BindingRequest)
+	if _, err := conn.Write(msg.Raw); err != nil {
+		return "", 0, fmt.Errorf("send stun request: %w", err)
+	}
 
-// 	buf := make([]byte, 1024)
-// 	n, err := conn.Read(buf)
-// 	if err != nil {
-// 		return "", 0, fmt.Errorf("读取响应失败：%s", err)
-// 	}
+	conn.SetDeadline(time.Now().Add(3 * time.Second))
+	defer conn.SetDeadline(time.Time{})
 
-// 	// 解析响应
-// 	var response stun.Message
-// 	response.Raw = buf[:n]
-// 	if err = response.Decode(); err != nil {
-// 		return "", 0, fmt.Errorf("解码stun失败%s", err)
-// 	}
+	buf := make([]byte, 1024)
+	n, err := conn.Read(buf)
+	if err != nil {
+		return "", 0, fmt.Errorf("read stun response: %w", err)
+	}
 
-// 	var xorAddr stun.XORMappedAddress
-// 	if err = xorAddr.GetFrom(&response); err != nil {
-// 		return "", 0, fmt.Errorf("获取映射地址失败: %s", err)
+	var response stun.Message
+	response.Raw = buf[:n]
+	if err = response.Decode(); err != nil {
+		return "", 0, fmt.Errorf("decode stun response: %w", err)
+	}
 
-// 	}
+	var xorAddr stun.XORMappedAddress
+	if err = xorAddr.GetFrom(&response); err != nil {
+		return "", 0, fmt.Errorf("get mapped address: %w", err)
+	}
 
-// 	return xorAddr.IP.String(), xorAddr.Port, nil
-// }
+	return xorAddr.IP.String(), xorAddr.Port, nil
+}
 
-// // 与STUN服务器握手UDP
-// func doUDPStunHandshake(conn *net.UDPConn, stunServerAddr *net.UDPAddr) (string, int, error) {
+// ================= TCP 保活 ========================
 
-// 	// 发送STUN请求
-// 	msg := stun.MustBuild(stun.TransactionID, stun.BindingRequest)
-// 	if _, err := conn.WriteToUDP(msg.Raw, stunServerAddr); err != nil {
-// 		return "", 0, fmt.Errorf("发送UDP STUN请求失败: %v", err)
-// 	}
+// STUN TCP 保活
+func tcpStunHealthKeepAlive(
+	ctx context.Context,
+	currentStunConn net.Conn,
+	publicIP string,
+	publicPort int,
+	localPort uint16,
+	localIP string,
+	onAlive func(),
+) error {
 
-// 	// 设置超时
-// 	conn.SetReadDeadline(time.Now().Add(3 * time.Second))
-// 	defer conn.SetReadDeadline(time.Time{})
+	// 进行第一次健康检查，同时保活
+	if !tcpStunHealthCheck(ctx, publicIP, publicPort) {
+		if ctx.Err() != nil {
+			return nil
+		}
+		return fmt.Errorf("tcp health check failed")
+	}
 
-// 	//读取响应
-// 	buf := make([]byte, 1024)
-// 	n, _, err := conn.ReadFromUDP(buf)
-// 	if err != nil {
-// 		return "", 0, fmt.Errorf("读取UDP响应失败: %v", err)
-// 	}
+	// 告诉外面TCP 检查ok
+	if onAlive != nil {
+		onAlive()
+	}
 
-// 	// 解析响应
-// 	var response stun.Message
-// 	response.Raw = buf[:n]
-// 	if err = response.Decode(); err != nil {
-// 		return "", 0, fmt.Errorf("解码UDP STUN失败%s", err)
-// 	}
+	// 开启28s 的检查
+	healthTicker := time.NewTicker(28 * time.Second)
+	defer healthTicker.Stop()
 
-// 	var xorAddr stun.XORMappedAddress
-// 	if err = xorAddr.GetFrom(&response); err != nil {
-// 		return "", 0, fmt.Errorf("获取UDP映射地址失败: %v", err)
+	for {
+		select {
+		case <-ctx.Done():
+			return nil
+		case <-healthTicker.C:
 
-// 	}
+			// 梯度健康检查
+			if tcpStunHealthCheck(ctx, publicIP, publicPort) {
+				continue
+			}
 
-// 	return xorAddr.IP.String(), xorAddr.Port, nil
-// }
+			// 梯度健康检查失败，先尝试用现有 stunConn 重新握手确认端口
+			_, port, err := doStunHandshake(currentStunConn)
+			if err == nil {
+				// 握手成功，但端口变了，说明 NAT 映射已变，无法恢复
+				if port != publicPort {
+					return fmt.Errorf("public port changed from %d to %d", publicPort, port)
+				}
+				// 端口未变，可能是网络抖动，继续等待
+				continue
+			}
 
-// // TCP首次健康检测保活
-// func firstTcpHealthKeep(publicIP string, expectedPublicPort int) bool {
-// 	sleepTime := 2 * time.Second
-// 	for i := 0; i < 3; i++ {
-// 		time.Sleep(sleepTime)
-// 		if tcpConnectCheck(publicIP, expectedPublicPort, 3*time.Second) {
-// 			return true
-// 		}
-// 		sleepTime = sleepTime * 2 // 休息2 4 8秒还不通就是死了
-// 	}
-// 	return false
-// }
+			// stunConn 本身也断了，尝试重新建连
+			currentStunConn.Close()
 
-// // TCP STUN 健康检测
-// func tcpStunHealthCheck(ctx context.Context, stunConn net.Conn, publicIP string, expectedPublicPort int, localPort uint16, service *model.Service) error {
-// 	healthTicker := time.NewTicker(28 * time.Second) // 每28s 检测一次
-// 	defer healthTicker.Stop()
+			stunServer, err := Runtime.STUNService.GetBestSTUNServer()
+			if err != nil {
+				stunServer, err = Runtime.STUNService.GetBackupSTUNServer()
+				if err != nil {
+					return fmt.Errorf("get stun server for reconnect: %w", err)
+				}
+			}
 
-// 	// 首次保活与检测
-// 	if !firstTcpHealthKeep(publicIP, expectedPublicPort) {
-// 		return fmt.Errorf("[%s] 首次保活失败，重启", service.Name)
-// 	}
-// 	service.PunchSuccess = true
+			localAddr := fmt.Sprintf("%s:%d", localIP, localPort)
+			newConn, dialErr := reuseport.Dial("tcp", localAddr, stunServer)
+			if dialErr != nil {
+				return fmt.Errorf("reconnect stun server %s failed: %w", stunServer, dialErr)
+			}
 
-// 	maxFailures := 3 // 失败阈值
-// 	failureCount := 0
-// 	currentStunConn := stunConn
+			_, newPort, handshakeErr := doStunHandshake(newConn)
+			if handshakeErr != nil {
+				newConn.Close()
+				return fmt.Errorf("stun handshake after reconnect failed: %w", handshakeErr)
+			}
 
-// 	logrus.Infof("[%s] 启动TCP健康检查 间隔28s", service.Name)
+			if newPort != publicPort {
+				newConn.Close()
+				return fmt.Errorf("public port changed from %d to %d", publicPort, newPort)
+			}
 
-// 	for {
-// 		select {
-// 		case <-ctx.Done():
-// 			return nil
-// 		case <-healthTicker.C:
-// 			// 策略1: 端到端服务检测+保活
-// 			if tcpConnectCheck(publicIP, expectedPublicPort, 3*time.Second) {
+			// 重连成功，端口一致，替换连接继续保活
+			currentStunConn = newConn
+		}
+	}
 
-// 				failureCount = 0 // 成功就重置
-// 				continue         // 服务正常，跳过stun检查
-// 			}
+}
 
-// 			failureCount++
+// 类似 Time.Sleep 但是开被ctx 终终端
+func sleepWithCtx(ctx context.Context, d time.Duration) bool {
+	select {
+	case <-time.After(d):
+		return true
+	case <-ctx.Done():
+		return false
+	}
+}
 
-// 			if failureCount == 1 { // 若果是第一次失败防止网络波动，来多一次
-// 				if tcpConnectCheck(publicIP, expectedPublicPort, 3*time.Second) {
-// 					failureCount = 0 // 成功就重置
-// 					continue
-// 				}
-// 			}
+// TCP STUN 梯度健康检查
+func tcpStunHealthCheck(ctx context.Context, publicIP string, expectedPublicPort int) bool {
+	// 第一次检查
+	if tcpConnectCheck(publicIP, expectedPublicPort, 3*time.Second) {
+		return true
+	}
 
-// 			logrus.Warnf("[%s] 端到端检查失败 (%d/%d)", service.Name, failureCount, maxFailures)
+	// 梯度检查1 3 5
+	sleepTime := 1 * time.Second
+	for i := 0; i < 3; i++ {
+		if !sleepWithCtx(ctx, sleepTime) {
+			return false
+		}
+		if tcpConnectCheck(publicIP, expectedPublicPort, 3*time.Second) {
+			return true
+		}
+		sleepTime += 2 * time.Second
+	}
+	return false
+}
 
-// 			// 策略2: STUN 检测NAT映射
-// 			_, port, err := doTcpStunHandshake(currentStunConn)
-// 			if err != nil {
-// 				logrus.Infof("STUN连接断开，尝试重连...")
+// TCP检查是否存活
+func tcpConnectCheck(host string, port int, timeout time.Duration) bool {
+	addr := fmt.Sprintf("%s:%d", host, port)
+	conn, err := net.DialTimeout("tcp", addr, timeout)
+	if err != nil {
+		return false
+	}
+	conn.Close()
+	return true
+}
 
-// 				// 关闭旧连接
-// 				currentStunConn.Close()
-// 				// 重连STUN
-// 				localAddr := fmt.Sprintf("%s:%d", global.StunConfig.LocalIP, localPort)
-// 				newConn, err := reuseport.Dial("tcp", localAddr, global.StunConfig.BestSTUN)
-// 				if err != nil {
-// 					return fmt.Errorf("STUN重连失败: %w", err)
-// 				}
+// ================= UDP 保活 ========================
 
-// 				// 验证新连接的端口
-// 				_, newPort, err := doTcpStunHandshake(newConn)
-// 				if err != nil {
-// 					newConn.Close()
-// 					return fmt.Errorf("重连后STUN验证失败: %w", err)
-// 				}
+// UDP STUN 保活
+func udpStunHealthKeepAlive(
+	ctx context.Context,
+	currentStunConn net.Conn,
+	// udpListener *net.UDPConn,
+	publicIP string,
+	publicPort int,
+	localPort uint16,
+	localIP string,
+	onAlive func(),
+) error {
+	// 先做一次外部连通性验证
+	if !udpStunHealthCheck(ctx, publicIP, publicPort, localPort) {
+		if ctx.Err() != nil {
+			return nil
+		}
+		return fmt.Errorf("udp health check failed")
+	}
+	if onAlive != nil {
+		onAlive()
+	}
 
-// 				// 检查端口是否变化
-// 				if newPort != expectedPublicPort {
-// 					newConn.Close()
-// 					return fmt.Errorf("公网端口漂移 %d -> %d", expectedPublicPort, newPort)
-// 				}
+	// UDP NAT 映射超时通常 30s，用 20s 心跳保住映射
+	keepAliveTicker := time.NewTicker(20 * time.Second)
+	// 外部连通性检查周期可以长一些
+	healthTicker := time.NewTicker(28 * time.Second)
+	defer keepAliveTicker.Stop()
+	defer healthTicker.Stop()
 
-// 				logrus.Infof("✅ STUN重连成功，端口保持 %d", newPort)
-// 				currentStunConn = newConn
-// 				continue
-// 			}
+	for {
+		select {
+		case <-ctx.Done():
+			return nil
 
-// 			if failureCount >= maxFailures {
-// 				return fmt.Errorf("[%s] 连续 %d 次检查失败，重新打洞", service.Name, maxFailures)
-// 			}
+		case <-keepAliveTicker.C:
+			// 向 STUN 服务器发心跳，维持 NAT 映射
+			msg := stun.MustBuild(stun.TransactionID, stun.BindingRequest)
+			currentStunConn.SetWriteDeadline(time.Now().Add(3 * time.Second))
+			currentStunConn.Write(msg.Raw)
+			currentStunConn.SetWriteDeadline(time.Time{})
 
-// 			// STUN正常但端口变化，触发重启
-// 			if port != expectedPublicPort {
-// 				return fmt.Errorf("❌ 公网端口漂移 %d -> %d，需要重新打洞", expectedPublicPort, port)
-// 			}
+		case <-healthTicker.C:
+			if udpStunHealthCheck(ctx, publicIP, publicPort, localPort) {
+				continue
+			}
 
-// 			// STUN正常但服务持续失败，可能是上游服务问题
-// 			logrus.Warnf("STUN端口正常但服务检查持续失败（已检查 %d 次），可能需要检查上游服务", maxFailures)
-// 		}
-// 	}
+			// 外部连通性失败，先用现有 conn 重新握手确认端口
+			_, port, err := doStunHandshake(currentStunConn)
+			if err == nil {
+				if port != publicPort {
+					return fmt.Errorf("public port changed from %d to %d", publicPort, port)
+				}
+				continue
+			}
 
-// }
+			// stunConn 断了，重新建连
+			currentStunConn.Close()
+			stunServer, err := Runtime.STUNService.GetBestSTUNServer()
+			if err != nil {
+				stunServer, err = Runtime.STUNService.GetBackupSTUNServer()
+				if err != nil {
+					return fmt.Errorf("get stun server for reconnect: %w", err)
+				}
+			}
 
-// // UDP健康检测
-// func udpStunHealthCheck(udpConn *net.UDPConn, stunServer *net.UDPAddr, expectedPublicPort int, localPort uint16) error {
-// 	healthTicker := time.NewTicker(280 * time.Second) // 每28s 健康检测一次
-// 	defer healthTicker.Stop()
+			localAddr := fmt.Sprintf("%s:%d", localIP, localPort)
+			newConn, dialErr := reuseport.Dial("udp", localAddr, stunServer)
+			if dialErr != nil {
+				return fmt.Errorf("reconnect stun server %s failed: %w", stunServer, dialErr)
+			}
 
-// 	consecutiveFailures := 0 // 连续失败计数器
-// 	maxFailures := 3         // 失败阈值
-// 	currentConn := udpConn
+			_, newPort, handshakeErr := doStunHandshake(newConn)
+			if handshakeErr != nil {
+				newConn.Close()
+				return fmt.Errorf("stun handshake after reconnect failed: %w", handshakeErr)
+			}
 
-// 	logrus.Info("启动UDP健康检查 间隔28s")
+			if newPort != publicPort {
+				newConn.Close()
+				return fmt.Errorf("public port changed from %d to %d", publicPort, newPort)
+			}
 
-// 	for range healthTicker.C {
-// 		// STUN 检测NAT映射
-// 		_, port, err := doUDPStunHandshake(currentConn, stunServer)
-// 		if err != nil {
-// 			consecutiveFailures++
-// 			logrus.Warnf("UDP STUN检查失败 (%d/%d): %v", consecutiveFailures, maxFailures, err)
+			currentStunConn = newConn
+		}
+	}
+}
 
-// 			// 达到失败阈值，尝试重建
-// 			if consecutiveFailures >= maxFailures {
-// 				logrus.Infof("UDP连接异常，尝试重建...")
-// 				// 重建
-// 				currentConn.Close()
-// 				localAddr := &net.UDPAddr{
-// 					IP:   net.ParseIP(global.StunConfig.LocalIP),
-// 					Port: int(localPort),
-// 				}
-// 				newConn, err := net.ListenUDP("udp", localAddr)
-// 				if err != nil {
-// 					return fmt.Errorf("UDP重建失败: %v", err)
-// 				}
+// UDP 梯度健康检查
+func udpStunHealthCheck(ctx context.Context, publicIP string, publicPort int, localPort uint16) bool {
+	if udpConnectCheck(publicIP, publicPort, localPort, 3*time.Second) {
+		return true
+	}
+	sleepTime := 1 * time.Second
+	for i := 0; i < 3; i++ {
+		if !sleepWithCtx(ctx, sleepTime) {
+			return false
+		}
+		if udpConnectCheck(publicIP, publicPort, localPort, 3*time.Second) {
+			return true
+		}
+		sleepTime += 2 * time.Second
+	}
+	return false
+}
 
-// 				// 验证新的连接，确保外部端口没变
-// 				_, newPort, err := doUDPStunHandshake(newConn, stunServer)
-// 				if err != nil {
-// 					newConn.Close()
-// 					return fmt.Errorf("重建后STUN验证失败: %v", err)
-// 				}
+// 向 publicIP:publicPort 发一个探测包，用临时 socket 收回包，不干扰 udpListener
+func udpConnectCheck(publicIP string, publicPort int, localPort uint16, timeout time.Duration) bool {
+	// 发探测包
+	dst := fmt.Sprintf("%s:%d", publicIP, publicPort)
+	probe, err := net.Dial("udp", dst)
+	if err != nil {
+		return false
+	}
+	defer probe.Close()
 
-// 				// 判断端口变没，如果变了必须退出，重新打洞
-// 				if newPort != expectedPublicPort {
-// 					newConn.Close()
-// 					return fmt.Errorf("公网端口漂移 %d -> %d", expectedPublicPort, newPort)
-// 				}
+	probe.SetDeadline(time.Now().Add(timeout))
+	if _, err = probe.Write([]byte("ping")); err != nil {
+		return false
+	}
 
-// 				logrus.Infof("✅ UDP重建成功，端口保持 %d", newPort)
-// 				currentConn = newConn
-// 				consecutiveFailures = 0 // 重建成功，重置计数
-// 			}
-// 			continue
-// 		}
+	// 单独开一个临时 socket 在同端口收回包，不碰 udpListener
+	tempConn, err := reuseport.ListenPacket("udp", fmt.Sprintf(":%d", localPort))
+	if err != nil {
+		return false
+	}
+	defer tempConn.Close()
 
-// 		// STUN 正常 但是端口变了，得重启
-// 		if port != expectedPublicPort {
-// 			return fmt.Errorf("❌ 公网端口变化 %d -> %d", expectedPublicPort, port)
-// 		}
-
-// 		// STUN正常
-// 		consecutiveFailures = 0
-// 	}
-// 	return nil
-
-// }
-
-// // tcpConnectCheck 通用 TCP 连通性检查
-// func tcpConnectCheck(host string, port int, timeout time.Duration) bool {
-// 	addr := fmt.Sprintf("%s:%d", host, port)
-// 	conn, err := net.DialTimeout("tcp", addr, timeout)
-// 	if err != nil {
-// 		logrus.Debugf("TCP连接检查失败 %s: %v", addr, err)
-// 		return false
-// 	}
-// 	conn.Close()
-// 	logrus.Debugf("TCP连接检查 OK %s", addr)
-// 	return true
-// }
+	tempConn.SetDeadline(time.Now().Add(timeout))
+	buf := make([]byte, 64)
+	_, _, err = tempConn.ReadFrom(buf)
+	return err == nil
+}
